Assign unique key to errors.
Such that an error assigns a `key` field that is unique for the field
derived from all parent properties and concatenated with a period.

The returned `fields` object now uses the unique key when collating the
array data such that there should no longer be collisions on error field
names.

Closes #33.